requirements/base.txt
awesome-slugify==1.6.5
certifi==2018.1.18
chardet==3.0.4
Django==2.0.4
djangorestframework==3.8.1
idna==2.6
pytz==2018.3
regex==2018.2.21
requests==2.18.4
six==1.11.0
slackclient==1.2.1
Unidecode==0.4.21
urllib3==1.22
websocket-client==0.47.0
pokebase==1.2.0
